+++ CHANGELOG.md
@@ -1,2 +1,12 @@
-# v0.1.0
+Change Log
+==========
 
+## v0.1.0
+
+- Initial release
+
+## v0.2.0
+
+- fixed multi-page queries when `--num N` was greater that the initial result size
+- added `--np` and `--noprompt` options to just search and exit
+- added `-l` `--language` command line options and `language` option in config file